feat(BAC-12):added the transcript.py file to routers and updated it
